Fix 64-bit build error
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@9279 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/MapReader.cpp
@@ -1127,7 +1127,7 @@ int CMapReader::ParseTerrain()
 		{	LOGERROR(L"CMapReader::ParseTerrain() failed to get '%hs' property", #prop);\
 			throw PSERROR_Game_World_MapLoadFailed("Error parsing terrain data.\nCheck application log for details"); }
 
-	size_t size;
+	u32 size;
 	GET_TERRAIN_PROPERTY(size, size)
 
 	m_PatchesPerSide = size / PATCH_SIZE;
